Move Plifal, Odeshin, and Grition out of the adapter folder
