+++ packages/open-schema-type-script/src/example/ciYamlFile/assertableCiYamlFileContents.ts
@@ -1,9 +1,9 @@
 import fs from 'fs';
 import yaml from 'yaml';
 import { Estinant2 } from '../../core/estinant';
-import { Grition } from '../../type-script-adapter/grition';
-import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
-import { Plifal } from '../../type-script-adapter/plifal';
+import { Grition } from '../custom-constructs/grition';
+import { Odeshin, ODESHIN_GEPP } from '../custom-constructs/odeshin';
+import { Plifal } from '../custom-constructs/plifal';
 import {
   ActualCiYamlFilePlifal,
   ACTUAL_CI_YAML_FILE_GEPPP,
