+++ packages/open-schema-type-script/src/example/ciYamlFile/assertableCiYamlFileContentsDifference.ts
@@ -1,9 +1,9 @@
 import assert from 'assert';
 import { OnamaEstinant } from '../../core/estinant';
 import { TropoignantTypeName } from '../../core/tropoignant';
-import { Grition } from '../../type-script-adapter/grition';
-import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
-import { Plifal } from '../../type-script-adapter/plifal';
+import { Grition } from '../custom-constructs/grition';
+import { Odeshin, ODESHIN_GEPP } from '../custom-constructs/odeshin';
+import { Plifal } from '../custom-constructs/plifal';
 import {
   AssertableCiYamlFileContentsOdeshin,
   ASSERTABLE_CI_YAML_FILE_CONTENTS_GEPPP,
