+++ packages/open-schema-type-script/src/example/ciYamlFile/expectedCiYamlFileContentsConfiguration.ts
@@ -1,6 +1,6 @@
-import { Grition } from '../../type-script-adapter/grition';
-import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
-import { Plifal } from '../../type-script-adapter/plifal';
+import { Grition } from '../custom-constructs/grition';
+import { Odeshin, ODESHIN_GEPP } from '../custom-constructs/odeshin';
+import { Plifal } from '../custom-constructs/plifal';
 import {
   CiYamlFileContents,
   CommentedSteps,
