+++ packages/open-schema-type-script/src/example/custom-constructs/plifal.ts
@@ -1,5 +1,5 @@
-import { GeppTuple } from '../core/gepp';
-import { Quirm } from '../core/quirm';
+import { GeppTuple } from '../../core/gepp';
+import { Quirm } from '../../core/quirm';
 import { Odeshin, OdeshinGepp, ODESHIN_GEPP } from './odeshin';
 
 /**
