+++ packages/open-schema-type-script/src/example/file/fileA.ts
@@ -8,9 +8,9 @@ import {
 } from '../../utilities/file/fileExtensionSuffixIdentifier';
 import { getFileMetadata } from '../../utilities/file/getFileMetadata';
 import { getNestedFilePaths } from '../../utilities/file/getNestedFilePaths';
-import { Grition } from '../../type-script-adapter/grition';
-import { Odeshin, ODESHIN_GEPP } from '../../type-script-adapter/odeshin';
-import { Plifal } from '../../type-script-adapter/plifal';
+import { Grition } from '../custom-constructs/grition';
+import { Odeshin, ODESHIN_GEPP } from '../custom-constructs/odeshin';
+import { Plifal } from '../custom-constructs/plifal';
 import {
   FileAConfigurationOdeshin,
   FILE_A_CONFIGURATION_GEPP,
